requirements.txt
pyserial==3.5
opencv-python
